backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
openai==1.84.0
requests==2.31.0
apscheduler==3.10.4 